% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_function.R
\name{census}
\alias{census}
\title{census}
\usage{
census(col_name)
}
\arguments{
\item{col_name}{species name column in original dataset}
}
\description{
This function allows you to quantify species representation in your dataset.
}
\examples{
census <- function(col_name)
}
\keyword{species}
