% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_index_function.R
\name{div_index}
\alias{div_index}
\title{div_index}
\usage{
div_index(species_col, pop_col)
}
\arguments{
\item{species_col}{species_name column from census() function output}

\item{pop_col}{individuals column from census() function output}
}
\description{
This function allows you explore your census data with several diversity indices (Species Richness, Simpson's Index, Shannon's Index, Berger-Parker Index) and evenness measures (Simpson's Evenness, Shannon's Evenness)
}
\examples{
div_index <- function(species_col, pop_col)
div_ind <- div_index(c$species_names, c$individuals)
}
\keyword{diversity}
